Refactoring para endpoints callback
