+++ back/routes/auth/github.js
@@ -36,37 +36,17 @@ router.get('/github/callback', passport.authenticate(strategy_name, {  session:f
     */
 
     console.log("New request GET to /github/callback");
-    console.log(req.user);
     const github_data = req.user._json;
     const user_id = req.query.state;
-    console.log(`state: ${user_id}`);
+    const provider_user_id = github_data.id;
+    const provider_email = github_data.email;
 
-    if (user_id){
-      console.log(`Connect the github account to the user ${user_id}`);
-      // TODO: create the relation between user and provider for user_id and provider(github_data)
-    }else{
-      console.log(`This is a login event. Check in the database if exists some user with this github account.
-        Login if exists, otherwise create a new user and connect with the github account`);
-      // TODO: Check if exists a user with this github account and log in him.
-      // TODO: If not exists, create the user and create the relation
-      //       between user and provider for user_id and provider(github_data)
-    }
+    const token = passport_callback(strategy_name, provider_user_id, provider_email, user_id);
 
-    const user = {id: 1, name: "Mauricio"};  // TODO: get the user data for the created or connected user
-
-    // TODO: generate a new token for login
-    const token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c";
-
-    const url_front = `http://localhost:5500/?token=${token}`;
+    const url_front = `${process.env.URL_FRONT}/?token=${token}`;
 
     res.redirect(301, url_front);
 
-    // let data = {
-    //   'success': true,
-    //   'message': `Authentication or connection successfully created for the user ${user_id}`,
-    //   'data': user
-    // }
-    // res.json(data)
   }
 );
