+++ back/sample.env
@@ -1,3 +1,4 @@
+URL_FRONT=
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
 GOOGLE_CALLBACK=
